tidy check

+++ src/function/scalar/string/overlay.cpp
@@ -97,7 +97,7 @@ void OverlayFunction(DataChunk &args, ExpressionState &state, Vector &result) {
 				result_validity.SetInvalid(i);
 				continue;
 			}
-			int64_t count_value = counts[count_idx] < 0 ? repls->GetSize() : counts[count_idx];
+			int64_t count_value = counts[count_idx] < 0 ? (int64_t)repls->GetSize() : counts[count_idx];
 			result_entries[i] =
 			    OverlayScalar(result, inputs[input_idx], repls[repl_idx], starts[start_idx], count_value);
 		}
